+++ zhurong-platform-core/src/main/java/com/zhurong/platform/core/clientimport/entity/ProductionOrder.java
@@ -22,7 +22,7 @@ public class ProductionOrder extends ClientImportRecordBase {
     private String matRef;
 
     @TableField("Thickness")
-    private BigDecimal thickness;
+    private Float thickness;
 
     @TableField("WrkRef")
     private String wrkRef;
